Create Project2.c
The goal here was "Several students want to share the same bathroom. If a woman is currently using it, any women at the front of the queue may enter, same for men. If the next person in line is of the opposite gender however, they must wait until the bathroom is empty. Use any synchronization primitives presented in our text."